modified:   FormValidator.py 	modified:   PattableQueries.py

+++ FormValidator.py
@@ -13,7 +13,7 @@ TODO:
 
 def RepresentsInt(s):
     try: 
-        int(s)
+        float(s)
         return True
     except ValueError:
         return False
